FREEPBX-9139 UCP doesnt create voicemail extensions directory for new extensions

+++ Voicemail.class.php
@@ -181,6 +181,9 @@ class Voicemail implements \BMO {
 		$file = $this->checkFileType($vmfolder, $type);
 		if(isset($this->greetings[$type]) && !empty($file)) {
 			foreach(glob($vmfolder."/".$type."*.*") as $filename) {
+				if(!file_exists($filename)) {
+					continue;
+				}
 				if(!unlink($filename)) {
 					return false;
 				}
@@ -200,6 +203,9 @@ class Voicemail implements \BMO {
 		$o = $this->getVoicemailBoxByExtension($ext);
 		$context = $o['vmcontext'];
 		$vmfolder = $this->vmPath . '/'.$context.'/'.$ext;
+		if(!file_exists($vmfolder)) {
+			mkdir($vmfolder);
+		}
 		if(isset($this->greetings[$source]) && isset($this->greetings[$target])) {
 			$tfile = $this->checkFileType($vmfolder, $target);
 			if(!empty($tfile)) {
@@ -224,6 +230,9 @@ class Voicemail implements \BMO {
 		$o = $this->getVoicemailBoxByExtension($ext);
 		$context = $o['vmcontext'];
 		$vmfolder = $this->vmPath . '/'.$context.'/'.$ext;
+		if(!file_exists($vmfolder)) {
+			mkdir($vmfolder);
+		}
 		if(isset($this->greetings[$type])) {
 			$file = $this->checkFileType($vmfolder, $type);
 			$tempf = $vmfolder . "/" . $type . "_tmp.".$format;
